// src/app/models/product.model.ts
export interface Product {
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[];
    isCompared?: boolean;
  }
  
  export interface ProductResponse {
    products: Product[];
    total: number;
    skip: number;
    limit: number;
  }